wip: feat: skemium orb
This commit introduces the Skemium orb that allows users of it easily
executing skemium checks.

An example usage can be seen in `.circleci/orb.yml`; currently the orb
only provides a single command that installs skemium, generates the
schema to ensure they're up-to-date, and then ensures they're
backward-compatible with the schemas on the target branch (`main`).

Users are still responsible for setting up the DB and populating it
with their schema, since this is something that isn't standardised.

This commit contains two different approaches to testing:
1) Through the `orb-tools` SDK in CircleCI: this does a basic "positive"
   check where we expect the pipeline to be green. However, because it's
   apparently not possible to check / invert exit codes of orb
   executions, I've also added
2) a "test suite" - a shell script - that executes the scripts of the
   orb with both positive and negative cases to ensure it behaves
   exactly like we expect to.

This commit still contains a few TODOs:
- documentation currently points to nowhere. We should write some docs
  to describe what users need to do to ensure schemas are up to date
  (but also, since this depends on the team's setups, I'm not sure how
  much we can actually write about that...)
- the orb isn't actually released yet. As far as I can tell someone with
  permissions will need to `circleci orb init` or `circleci orb publish`
  to register the orb at CircleCI's orb registry. Then the steps in the
  `.circleci/config.yml` and `.circleci/orb.yml` can be uncommented and
  updated to actually push the orb.

There's a few other `TODOs` which are either ideas / extensions for the
future, or open questions I might have. Please have a good luck at
those :-)